Initial Project Done, Next Step Auth

+++ UserManager/views.py
@@ -3,6 +3,7 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.hashers import make_password
 
 
+from django.http import Http404
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -21,11 +22,11 @@ class SaveUserInformation(APIView):
         if serializer.is_valid():
             user_info= models.UserInformation.objects.create(
                 user = User.objects.get(id=request.data["user"]),
-                bloodType = models.BloodType.objects.get(bloodType=request.data["bloodType"]),
+                bloodType = models.BloodType.objects.get(bloodtype=request.data["bloodType"]),
                 name = request.data["name"],
                 mobile = request.data["mobile"],
                 govern = models.Govern.objects.get(name=request.data["govern"]),
-                state = models.GovernStates.objects.get(name=request.data["state"]),
+                state = models.GovernState.objects.get(name=request.data["state"]),
                 Card_number = request.data["Card_number"],
                 date_of_donation = request.data["date_of_donation"],
                 age = request.data["age"],
@@ -36,11 +37,22 @@ class SaveUserInformation(APIView):
         return Response(data=serializer.errors, status=status.HTTP_406_NOT_ACCEPTABLE)
 
 class UpdateInformation(APIView):
+    def get_user(self,request):
+        try:
+            user = User.objects.get(id=request.data["user"])
+            user_info = models.UserInformation.objects.get(user=user)
+            return user_info
+
+        except ObjectDoesNotExist:
+
+            raise Http404
+            
 
     def get(self, request):
         try:
             user = User.objects.get(id=request.data["user"])
-            serializer = UserSerializer(instance=user, many=False)
+            user_info = models.UserInformation.objects.get(user=user)
+            serializer = UserInformationSerializer(instance=user_info, many=False)
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         except ObjectDoesNotExist:
 
@@ -48,13 +60,14 @@ class UpdateInformation(APIView):
 
 
     def put(self, request):
-        try:
-            user = User.objects.get(id=request.data["user"])
-            user_info = models.UserInformation.objects.get(user=user)
-        except ObjectDoesNotExist:
-            return Response(data={"message": "NOT FOUND!"}, status=status.HTTP_404_NOT_FOUND)
-            
-        if user.check_password(request.data["password"]):
+        # try:
+        #     user = User.objects.get(id=request.data["user"])
+        #     user_info = models.UserInformation.objects.get(user=user)
+        # except ObjectDoesNotExist:
+        #     return Response(data={"message": "NOT FOUND!"}, status=status.HTTP_404_NOT_FOUND)
+        user_info = self.get_user(request=request)
+
+        if user_info.user.check_password(request.data["password"]):
             serializer = UserInformationSerializer(instance=user_info, data=request.data)
             if serializer.is_valid():
                 serializer.save(password=make_password(request.data["password"]))
@@ -68,15 +81,24 @@ class UpdateInformation(APIView):
 
         
 class DeleteUser(APIView):
-
-    def delete(self, request):
+    def get_user(self,request):
         try:
             user = User.objects.get(id=request.data["user"])
-        
+            return user
         except ObjectDoesNotExist:
 
-            return Response(data={"message": "NOT FOUND!"}, status=status.HTTP_404_NOT_FOUND) 
+            raise Http404
+            
 
-        user.delete()
-        return Response(data={"messgae": "Account Deleted"}, status=status.HTTP_204_NO_CONTENT)
+    def delete(self, request):
+        # try:
+        #     user = User.objects.get(id=request.data["user"])
+        
+        # except ObjectDoesNotExist:
 
+        #     return Response(data={"message": "NOT FOUND!"}, status=status.HTTP_404_NOT_FOUND) 
+        user = self.get_user(request=request)
+        if user.check_password(request.data["password"]):
+            user.delete()
+            return Response(data={"messgae": "Account Deleted"}, status=status.HTTP_204_NO_CONTENT)
+        return Response(data={"message":"Wrong Password!"}, status=status.HTTP_204_NO_CONTENT)
